Fix MinGW compiler warnings
reg. comparison of integer expressions of different signedness: 'size_t' {aka 'long long unsigned int'} and 'int' (-Wsign-compare)

+++ src/engraving/tests/testbase.cpp
@@ -207,7 +207,7 @@ bool MTest::saveMimeData(QByteArray mimeData, const QString& saveName)
     }
 
     size_t size = f.write(mimeData);
-    return size == mimeData.size();
+    return size == static_cast<size_t>(mimeData.size());
 }
 
 //---------------------------------------------------------
